Move configuration into module
